Split backend progress report ownership
